@@ -2,23 +2,27 @@
   "name": "catstronauts-client",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "description": "front-end demo app for Apollo's lift-off course",
   "dependencies": {
+    "@apollo/client": "^3.8.10",
     "@apollo/space-kit": "^9.3.1",
     "@emotion/cache": "^11.4.0",
     "@emotion/core": "^10.1.1",
-    "@emotion/react": "^11.4.0",
+    "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.3.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-markdown": "^6.0.2",
-    "react-player": "^2.6.0",
-    "react-router-dom": "^6.8.0"
+    "graphql": "^16.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^9.0.1",
+    "react-player": "^2.14.1",
+    "react-router-dom": "^6.21.3"
   },
   "scripts": {
     "test": "vitest",
     "start": "vite",
-    "build": "vite build"
+    "build": "vite build",
+    "generate": "graphql-codegen"
   },
   "browserslist": {
     "production": [
@@ -33,21 +37,23 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^29.2.6",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
+    "@graphql-codegen/cli": "^5.0.0",
+    "@graphql-codegen/client-preset": "^4.1.0",
+    "@testing-library/jest-dom": "^6.3.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.11.6",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
     "@types/react-router-dom": "^5.3.3",
-    "typescript": "^5.1.3",
-    "@vitejs/plugin-react": "^4.0.0",
-    "@vitest/ui": "^0.32.2",
-    "happy-dom": "^9.20.3",
-    "jest": "^27.5.1",
-    "vite": "^4.3.9",
-    "vitest": "^0.32.2"
+    "@vitejs/plugin-react": "^4.2.1",
+    "@vitest/ui": "^1.2.1",
+    "happy-dom": "^13.3.1",
+    "jest": "^29.7.0",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.12",
+    "vitest": "^1.2.1"
   },
   "overrides": {
     "nth-check": "2.0.1"
