@@ -6,15 +6,27 @@
   "scripts": {
     "compile": "tsc",
     "dev": "ts-node-dev --respawn ./src/index.ts",
-    "start": "npm run compile && nodemon ./dist/index.js"
+    "start": "npm run compile && nodemon ./dist/index.js",
+    "generate": "graphql-codegen"
   },
   "devDependencies": {
-    "dotenv": "^8.6.0",
-    "nodemon": "^2.0.20",
+    "@graphql-codegen/cli": "^5.0.0",
+    "@graphql-codegen/typescript": "^4.0.1",
+    "@graphql-codegen/typescript-resolvers": "^4.0.1",
+    "dotenv": "^16.4.1",
+    "nodemon": "^3.0.3",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.1.3"
+    "typescript": "^5.3.3"
   },
   "author": "Raphael Terrier @R4ph-t",
   "license": "MIT",
-  "private": true
+  "private": true,
+  "dependencies": {
+    "@apollo/datasource-rest": "^6.2.2",
+    "@apollo/server": "^4.10.0",
+    "@graphql-tools/mock": "^9.0.1",
+    "@graphql-tools/schema": "^10.0.2",
+    "graphql": "^16.8.1",
+    "graphql-tag": "^2.12.6"
+  }
 }
