@@ -1,12 +1,18 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import GlobalStyles from './styles';
 import Pages from './pages';
+import { createRoot } from 'react-dom/client';
+import { ApolloProvider } from '@apollo/client';
+import { client } from './lib/apollo-client';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
-    <GlobalStyles />
-    <Pages />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <ApolloProvider client={client}>
+      <GlobalStyles />
+      <Pages />
+    </ApolloProvider>
+  </React.StrictMode>
 );
