@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from '@emotion/styled';
-// import { Link } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { colors, IconArrowRight, IconDoubleArrowRight } from '../styles';
 import { humanReadableTimeFromSeconds } from '../utils/helpers';
 
@@ -8,25 +8,24 @@ import { humanReadableTimeFromSeconds } from '../utils/helpers';
  * Module Navigation: displays a list of modules titles
  * from a track and navigates to the modules page
  */
-const ModulesNav: React.FC<{module: any, track: any}> = ({ module, track }) => {
+const ModulesNav: React.FC<{ module: any; track: any }> = ({
+  module,
+  track,
+}) => {
   return (
     <ModulesNavContainer>
       <ModuleTitle>
         <h4>
-          {/* Need to comment this out until Link is placed within a Router */}
-          {/* <Link to="../.."> */}
-          {track.title}
-          {/*</Link> */}
+          <Link to="../..">{track.title}</Link>
         </h4>
       </ModuleTitle>
       <ModulesList>
         {track.modules.map((navModule: any) => (
           <ModuleListItem key={`module_${navModule.id}`}>
             <div>
-              {/* Need to comment this out until Link is placed within a Router */}
-              {/* <ModuleNavStyledLink
+              <ModuleNavStyledLink
                 to={`/track/${track.id}/module/${navModule.id}`}
-              > */}
+              >
                 <ModuleListItemContent isActive={navModule.id === module.id}>
                   {navModule.id === module.id ? (
                     <IconDoubleArrowRight width="14px" />
@@ -36,7 +35,7 @@ const ModulesNav: React.FC<{module: any, track: any}> = ({ module, track }) => {
                   <div>{navModule.title}</div>
                   <div>{humanReadableTimeFromSeconds(navModule.length)}</div>
                 </ModuleListItemContent>
-              {/* </ModuleNavStyledLink> */}
+              </ModuleNavStyledLink>
             </div>
           </ModuleListItem>
         ))}
@@ -96,11 +95,11 @@ const ModuleListItem = styled.li((props) => ({
   },
 }));
 
-// const ModuleNavStyledLink = styled(Link)({
-//   textDecoration: 'none',
-//   display: 'flex',
-//   alignItems: 'center',
-// });
+const ModuleNavStyledLink = styled(Link)({
+  textDecoration: 'none',
+  display: 'flex',
+  alignItems: 'center',
+});
 
 const ModuleListItemContent = styled.div((props: { isActive: boolean }) => ({
   backgroundColor: props.isActive ? colors.black.base : colors.black.light,
