@@ -1,7 +1,10 @@
 import { useState, useEffect } from 'react';
 
 function getWindowDimensions() {
-  const { innerWidth: width, innerHeight: height }: { innerWidth: number, innerHeight: number} = window;
+  const {
+    innerWidth: width,
+    innerHeight: height,
+  }: { innerWidth: number; innerHeight: number } = window;
   return {
     width,
     height,
